new upgrade

+++ construction_factory_bot/database/models.py
@@ -161,7 +161,7 @@ class Employee(Base):
     __tablename__ = "employees"
     
     id = Column(Integer, primary_key=True, index=True)
-    telegram_id = Column(BigInteger, unique=True, nullable=True)
+    telegram_id = Column(Integer, unique=True, nullable=True)
     full_name = Column(String(100), nullable=False)
     phone_number = Column(String(20), nullable=False)
     position = Column(String(50), nullable=False)
